Commenting code (util)

+++ src/main/java/edu/chalmers/vaporwave/util/ErrorMessage.java
@@ -7,15 +7,19 @@ import javafx.scene.image.ImageView;
 import java.util.Timer;
 import java.util.TimerTask;
 
+/**
+ * This class shows a simple error window if any unchecked exceptions occur.
+ */
 public class ErrorMessage {
     private static Group root;
     private static ImageView imageView;
 
-    private static long timeInSeconds;
+    private static int timeInSeconds;
     private static Timer timer;
 
     private static boolean shown;
 
+    // First initiated in Main class
     public static void init(Group rootIn) {
         root = rootIn;
 
@@ -29,6 +33,8 @@ public class ErrorMessage {
         shown = false;
     }
 
+    // Except showing the message, this also prints the exceptions stacktrace and then
+    // terminates the program after a set amount of seconds
     public static void show(Exception e) {
         if (!shown) {
             e.printStackTrace();
